Public Notebook Delete Pop Up function working (2-5)

+++ application/controllers/PublicNotebook.php
@@ -141,7 +141,7 @@
                 {
                     redirect('publicnotebook');
                 }
-                else
+                else if($action == 'YES')
                 {  
                     $this->deletePublicPage();
                 }
